src/requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==5.0.3
distlib==0.3.0
filelock==3.0.12
Flask==1.1.1
Flask-Caching==1.8.0
Flask-Limiter==1.1.0
Flask-RESTful==0.3.8
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
limits==1.5
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
peewee==3.13.1
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.5
python-dotenv==0.11.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
stringcase==1.2.0
toml==0.10.0
tqdm==4.43.0
urllib3==1.25.8
virtualenv==20.0.5
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.0.0
